fix: code cleanup and minor module refactor, using RAII and macro for proper CUDA calls
