+++ wayland-display-core/src/utils/video_info.rs
@@ -1,10 +1,10 @@
-use crate::utils::allocator::gst_cuda_ffi;
+use crate::utils::allocator::cuda;
 use gst_video::{VideoInfo, VideoInfoDmaDrm};
 
 #[derive(Debug, Clone)]
 pub struct CUDAParams {
     pub video_info: VideoInfoDmaDrm,
-    pub cuda_context: gst_cuda_ffi::CUDAContext,
+    pub cuda_context: cuda::CUDAContext,
 }
 
 #[derive(Debug, Clone)]
